// libs/xplajs/src/ibc/core/channel/v1/channel.d.ts
export interface PacketId {
    /** channel port identifier */
    portId: string;
    /** channel unique identifier */
    channelId: string;
    /** packet sequence */
    sequence: bigint;
}
export interface PacketIdProtoMsg {
    typeUrl: "/ibc.core.channel.v1.PacketId";
    value: Uint8Array;
}
/**
 * PacketId is an identifer for a unique Packet
 * Source chains refer to packets by source port/channel
 * Destination chains refer to packets by destination port/channel
 */
export interface PacketIdAmino {
    /** channel port identifier */
    port_id: string;
    /** channel unique identifier */
    channel_id: string;
    /** packet sequence */
    sequence: string;
}
export interface PacketIdAminoMsg {
    type: "cosmos-sdk/PacketId";
    value: PacketIdAmino;
}
// ...
export declare const PacketId: {
    typeUrl: string;
    aminoType: string;
    is(o: any): o is PacketId;
    isAmino(o: any): o is PacketIdAmino;
    encode(message: PacketId, writer?: BinaryWriter): BinaryWriter;
    decode(input: BinaryReader | Uint8Array, length?: number): PacketId;
    fromPartial(object: DeepPartial<PacketId>): PacketId;
    fromAmino(object: PacketIdAmino): PacketId;
    toAmino(message: PacketId): PacketIdAmino;
    fromAminoMsg(object: PacketIdAminoMsg): PacketId;
    toAminoMsg(message: PacketId): PacketIdAminoMsg;
    fromProtoMsg(message: PacketIdProtoMsg): PacketId;
    toProto(message: PacketId): Uint8Array;
    toProtoMsg(message: PacketId): PacketIdProtoMsg;
};